tidy - remove test files + rename dps_helper_async to dps_helper
