+++ lcpdelta_python_package/src/lcp_delta/enact/dps_helper.py
@@ -1,19 +1,19 @@
-import time as pytime
+from collections.abc import Awaitable
+import logging
 import pandas as pd
 import threading
-import queue
 import inspect
 import asyncio
 import zoneinfo
 from datetime import datetime as dt
 from datetime import timezone, time, timedelta
 from functools import partial
-from typing import Callable
-from concurrent.futures import ThreadPoolExecutor
-from signalrcore.hub_connection_builder import HubConnectionBuilder
+from typing import Any, Callable
+from pysignalr.client import SignalRClient
 from lcp_delta.global_helpers import is_list_of_strings_or_empty, is_2d_list_of_strings
 from lcp_delta.enact.api_helper import APIHelper
 from lcp_delta.common.http.exceptions import EnactApiError
+import asyncio
 
 EPEX_SUBSCRIPTION_ID = "EPEX_TRADES"
 
@@ -32,141 +32,160 @@ class DPSHelper:
         max_workers:
             Maximum number of callback tasks that can run at once (only used when async_mode=True)
         """
+        # suppress logging pushes with no registered handlers
+        logging.getLogger("pysignalr").setLevel(logging.ERROR) 
+
         self.api_helper = APIHelper(username, public_api_key)
         self._multi_series_subscriptions: list[tuple[str, Callable, bool]] = []
         self._single_series_subscriptions: list[tuple[str, str]] = []
-        self._suppress_restart = False
+        self._connection_open = False
         self.last_updated_header = "DateTimeLastUpdated"
 
         self.async_mode = async_mode
         self.max_workers = max_workers
+
         if self.async_mode:
-            self._executor = ThreadPoolExecutor(max_workers=max_workers, thread_name_prefix="dps-callback")
-            self._series_queues: dict[str, queue.Queue] = {}
-            self._series_running: set[str] = set()
+            self._series_queues: dict[str, asyncio.Queue] = {}
+            self._series_semaphore = asyncio.Semaphore(max_workers)
             self._series_lock = threading.Lock()
+            self._series_running: set[str] = set()
 
-        self._initialise()
-        self.start_connection_monitor()
+        # Background event loop for SignalR
+        self._loop = None     
+        self._loop_ready = threading.Event()
+        self._client_initialised = threading.Event()
 
-    def _initialise(self):
+
+        # Start event loop on background thread
+        self._thread = threading.Thread(target=self._start_loop, daemon=True)
+        self._thread.start()
+
+        # init SignalR client in background loop
+        self._loop_ready.wait()
+        self._run_in_loop(self._initialise())
+
+
+    # Background event loop helpers
+    def _start_loop(self):
+        """Run a dedicated asyncio event loop in a background thread."""
+        self._loop = asyncio.new_event_loop()
+        asyncio.set_event_loop(self._loop)  
+        self._loop_ready.set()     
+        self._loop.run_forever()
+            
+    def _run_in_loop(self, coro):
+        """Submit a coroutine to the background loop from any thread."""
+        if not self._loop:
+            raise RuntimeError("Event loop not ready")
+        return asyncio.run_coroutine_threadsafe(coro, self._loop)
+    
+    def _ensure_async(self, func: Callable[..., Any]) -> Callable[..., Awaitable[Any]]:
+        if inspect.iscoroutinefunction(func):
+            return func
+
+        async def async_wrapper(*args, **kwargs):
+            return await asyncio.to_thread(func, *args, **kwargs)
+        return async_wrapper
+    
+    
+    # SignalR client initialization
+    async def _initialise(self):
         self.enact_credentials = self.api_helper.credentials_holder
-        self.data_by_single_series_subscription_id: dict[object, tuple[Callable[[pd.DataFrame], None], pd.DataFrame, bool]] = {}
-        access_token_factory = partial(self._fetch_bearer_token)
+        self.data_by_single_series_subscription_id: dict[
+            object, tuple[Callable[[pd.DataFrame], None], pd.DataFrame, bool]
+        ] = {}
 
-        local_url = "https://localhost:44330/dataHub"
-        staging_url = "https://enact-signalrhub-staging.azurewebsites.net/dataHub"
-        prod_url = self.api_helper.endpoints.DPS
-        self.hub_connection = (
-            HubConnectionBuilder()
-            .with_url(
-                prod_url,
-                options={"access_token_factory": access_token_factory},                 
-            )
-            .build()
+        access_token_factory = partial(self._fetch_bearer_token)
+        url = self.api_helper.endpoints.DPS
+
+        self.client = SignalRClient(
+            url,
+            access_token_factory=access_token_factory,
+            headers={"Authorization": f"Bearer {access_token_factory()}"},
         )
 
-        self.hub_connection.on_open(self._on_open)
-        self.hub_connection.on_close(self._on_close)
-        self.hub_connection.on_error(lambda e: print("SignalR error:", e))
-        self.hub_connection.on_reconnect(lambda: print("reconnected!"))
+        self.client.on_open(self._on_open)
+        self.client.on_close(self._on_close)
+        self.client.on_error(lambda e: print(f"SignalR error: {e}"))
 
-        success = self.hub_connection.start()
-        pytime.sleep(1)
-
-        if not success:
-            raise ValueError("connection failed")
+        # Start SignalR client
+        self._signalr_task = asyncio.create_task(self.client.run())
+        self._client_initialised.set()
 
     def _fetch_bearer_token(self):
         self.enact_credentials.get_bearer_token()
         return self.enact_credentials.bearer_token
+     
+    async def _add_subscription(self, request_object: list[dict[str, str]], subscription_id: str):
+        # Create wrapper to capture subscription_id in callback
+        async def on_JoinEnactPush(m):
+            result = m.result
+            await self._callback_received(result, subscription_id)
 
-    def _add_subscription(self, request_object: list[dict[str, str]], subscription_id: str):
-        self.hub_connection.send(
-            "JoinEnactPush", request_object, lambda m: self._callback_received(m.result, subscription_id)
-        )
+        await self.client.send(
+            "JoinEnactPush", 
+            request_object,
+            on_JoinEnactPush
+            )
 
-    def _on_open(self):
-        print("connection opened and handshake received ready to send messages")
+    async def _on_open(self):
+        print("Connection opened and handshake received ready to send messages")
         # Reconnect to prior pushes without increasing API usage
-        for push_group_name, subscription_id in self._single_series_subscriptions:
+        for push_group_name, _ in self._single_series_subscriptions:
             try:
-                self.hub_connection.send(
+                await self.client.send(
                     "ReconnectToPush",
                     [push_group_name],
-                    lambda m, sid=subscription_id: self._callback_received(m.result, sid, is_for_reconnect=True)
                 )
             except Exception as e:
                 print(f"Resubscribe failed (single): {e}")
 
-        for push_group_name, handler, parse_datetimes in self._multi_series_subscriptions:
+        for push_group_name, _, _ in self._multi_series_subscriptions:
             try:
-                self.hub_connection.send(
+                await self.client.send(
                     "ReconnectToPush",
                     [push_group_name],
-                    lambda m, h=handler, pdts=parse_datetimes: self._callback_received_multi_series(m.result, h, pdts, is_for_reconnect=True),
                 )
             except Exception as e:
                 print(f"Resubscribe failed (multi): {e}")
 
-    def _on_close(self):
+        self._connection_open = True
+
+    async def _on_close(self):
+        if(self._connection_open == False):
+            return       
+        
         print("Connection closed")
-        if self._suppress_restart:
-            return
+        self._connection_open = False
 
-        if not self.hub_connection.transport.is_running():
-            print("Attempting to restart hub connection")
-            self._restart_connection()
+    async def _add_multi_series_subscription(self, request_object: list, handle_data_method, parse_datetimes):
+        # Create wrapper to capture handle_data_method, parse_datetimes in callback
+        async def on_JoinMultiSeries(m):
+            result = m.result
+            await self._callback_received_multi_series(result, handle_data_method, parse_datetimes)
 
-    def _restart_connection(self):
-        try:
-            self.hub_connection.stop()
-        except Exception as e:
-            print(f"Error during stop: {e}")
-
-        pytime.sleep(1)
-        self._initialise()
-
-    def is_connection_alive(self):
-        return self.hub_connection and self.hub_connection.transport and self.hub_connection.transport.is_running()
-
-    def start_connection_monitor(self, check_interval_seconds=60):
-        if hasattr(self, "_monitor_thread") and self._monitor_thread.is_alive():
-            # Monitor is already running
-            return
-
-        self._stop_event = getattr(self, "_stop_event", threading.Event())
-
-        def monitor_loop():
-            backoff = 1
-            while not self._stop_event.wait(check_interval_seconds * backoff):
-                try:
-                    if not self.is_connection_alive():
-                        print("Connection not alive. Restarting...")
-                        self._restart_connection()
-                        backoff = min(backoff * 2, 8)
-                    else:
-                        backoff = 1
-                except Exception as e:
-                    print(f"Error during connection check: {e}")
-
-        self._monitor_thread = threading.Thread(target=monitor_loop, daemon=True)
-        self._monitor_thread.start()
-
-    def _add_multi_series_subscription(self, request_object: list, handle_data_method, parse_datetimes):
-        self.hub_connection.send(
+        await self.client.send(
             "JoinMultiSeries",
             request_object,
-            lambda m: self._callback_received_multi_series(m.result, handle_data_method, parse_datetimes),
+            on_JoinMultiSeries,
         )
 
     def subscribe_to_notifications(self, handle_notification_method: Callable[[object], None]):
-        self.hub_connection.send(
+        self._client_initialised.wait()
+        handle_notification_method = self._ensure_async(handle_notification_method)
+        self._run_in_loop(
+            self._add_notification_subscription(handle_notification_method)
+        )
+
+    async def _add_notification_subscription(self, handle_notification_method):
+        async def on_join_parent_company_notification_push(m):         
+            push_name = m.result["data"]["pushName"]
+            self.client.on(push_name, handle_notification_method)
+
+        await self.client.send(
             "JoinParentCompanyNotificationPush",
             [],
-            on_invocation=lambda m: self.hub_connection.on(
-                m.result["data"]["pushName"], lambda x: handle_notification_method(x)
-            ),
+            on_join_parent_company_notification_push,
         )
 
     def _initialise_series_subscription_data(
@@ -194,13 +213,18 @@ class DPSHelper:
             parse_datetimes,
         )
 
-    def _callback_received(self, m, subscription_id: str, is_for_reconnect: bool = False):
+    async def _callback_received(self, m, subscription_id: str, is_for_reconnect: bool = False):
         push_name = m["data"]["pushName"]
         if not is_for_reconnect:
             self._single_series_subscriptions.append((push_name, subscription_id))
-        self.hub_connection.on(push_name, lambda x: self._process_push_data(x, subscription_id))
 
-    def _callback_received_multi_series(self, m, handle_data_method, parse_datetimes, is_for_reconnect: bool = False):
+        # Create wrapper to capture subscription_id in callback
+        async def push_handler(x):
+            await self._process_push_data(x, subscription_id)
+        
+        self.client.on(push_name, push_handler)
+
+    async def _callback_received_multi_series(self, m, handle_data_method, parse_datetimes, is_for_reconnect: bool = False):
         if "messages" in m and len(m["messages"]) > 0:
             error_return = m["messages"][0]
             raise EnactApiError(error_return["errorCode"], error_return["message"], m)
@@ -208,12 +232,15 @@ class DPSHelper:
         push_name = m["data"]["pushName"]
         if not is_for_reconnect:
             self._multi_series_subscriptions.append((push_name, handle_data_method, parse_datetimes))
-        self.hub_connection.on(push_name, lambda x: self._process_multi_series_push(x, handle_data_method, parse_datetimes))
 
-    def _process_push_data(self, data_push, subscription_id):
+        async def push_handler(x):
+            await self._process_multi_series_push(x, handle_data_method, parse_datetimes)
 
+        self.client.on(push_name, push_handler)
+
+    async def _process_push_data(self, data_push, subscription_id):
         if subscription_id == EPEX_SUBSCRIPTION_ID:
-            self._enqueue_or_call(subscription_id, self.epex_trade_call_back, data_push)
+            await self._enqueue_or_call(subscription_id, self.epex_trade_call_back, data_push)
             return
 
         (user_callback, all_data, parse_datetimes) = self.data_by_single_series_subscription_id[subscription_id]
@@ -223,7 +250,7 @@ class DPSHelper:
         except Exception:
             series_id = str(subscription_id)
 
-        def work(data):
+        async def work(data):
             updated_data = self._handle_new_series_data(data, data_push, parse_datetimes)
             if updated_data is not data:
                 self.data_by_single_series_subscription_id[subscription_id] = (
@@ -232,76 +259,52 @@ class DPSHelper:
                     parse_datetimes,
                 )
 
-            self._invoke_handler(user_callback, updated_data)
+            await user_callback(updated_data)
 
-        self._enqueue_or_call(series_id, work, all_data)
+        await self._enqueue_or_call(series_id, work, all_data)
 
-    def _process_multi_series_push(self, data_push, handle_data_method, parse_datetimes):
+    async def _process_multi_series_push(self, data_push, handle_data_method, parse_datetimes):
         updated_data = self._handle_new_multi_series_data(data_push, parse_datetimes)
         series_id = data_push[0]["data"]["id"]
-        self._enqueue_or_call(series_id, handle_data_method, updated_data)
+        await self._enqueue_or_call(series_id, handle_data_method, updated_data)
 
-    def _enqueue_or_call(self, series_id: str, handler: Callable, data):
+    async def _enqueue_or_call(self, series_id: str, handler: Callable, data):
         """
         In blocking mode: call directly.
-        In async mode: enqueue (series-ordered) and schedule via shared thread pool respecting max_workers.
+        In async mode: enqueue (series-ordered) and schedule via shared event loop with max_workers limit.
         """
         if not self.async_mode:
-            self._invoke_handler(handler, data)
+            await handler(data)
             return
 
         first = False
         with self._series_lock:
             q = self._series_queues.get(series_id)
             if q is None:
-                q = self._series_queues[series_id] = queue.Queue()
-            q.put((handler, data))
+                q = self._series_queues[series_id] = asyncio.Queue()
+            q.put_nowait((handler, data))
             if series_id not in self._series_running:
                 self._series_running.add(series_id)
-                first = True  # we need to start the drain
+                first = True
 
         if first:
-            self._executor.submit(self._drain_one, series_id)
+            asyncio.create_task(self._drain_one(series_id))
 
-    def _drain_one(self, series_id: str):
-        """Process exactly one item from the series queue, then reschedule if needed."""
-        with self._series_lock:
-            q = self._series_queues.get(series_id)
-            if q is None or q.empty():
-                self._series_running.discard(series_id)
-                return
+    async def _drain_one(self, series_id: str):
+        """Process all items from the series queue sequentially, respecting max_workers limit."""
+        async with self._series_semaphore:
+            while True:
+                with self._series_lock:
+                    q = self._series_queues.get(series_id)
+                    if q is None or q.empty():
+                        self._series_running.discard(series_id)
+                        return
+                    handler, data = q.get_nowait()
 
-            handler, data = q.get_nowait()
-
-        try:
-            # Run the callback safely
-            self._invoke_handler(handler, data)
-        except Exception as e:
-            print(f"Error in callback for series {series_id}: {e}")
-        finally:
-            q.task_done()
-
-        # Reschedule the next item if there’s still work to do
-        with self._series_lock:
-            if not q.empty():
-                self._executor.submit(self._drain_one, series_id)
-            else:
-                self._series_running.discard(series_id)
-
-    def _invoke_handler(self, handler: Callable, data):
-        """
-        Runs sync or async handlers appropriately.
-        (Pool threads are not running an event loop, so we create one for async handlers.)
-        """
-        if inspect.iscoroutinefunction(handler):
-            try:
-                loop = asyncio.get_running_loop()
-                future = asyncio.run_coroutine_threadsafe(handler(data), loop)
-                future.result()
-            except RuntimeError:
-                asyncio.run(handler(data))
-        else:
-            handler(data)
+                try:
+                    await handler(data)
+                except Exception as e:
+                    print(f"Error in callback for series {series_id}: {e}")
 
     def _handle_new_series_data(
         self, all_data: pd.DataFrame, data_push_holder: list, parse_datetimes: bool
@@ -369,25 +372,70 @@ class DPSHelper:
             return df_return
         except Exception:
             return pd.DataFrame()
-
-    def terminate_hub_connection(self):
-        self.shutdown()
-
-    def shutdown(self):
-        self._suppress_restart = True
-        self._stop_event.set()
-
-        if self.hub_connection:
+        
+    
+    async def _cancel_signalr_task(self):
+        """Cancel the SignalR task gracefully"""
+        if hasattr(self, '_signalr_task') and not self._signalr_task.done():
+            self._signalr_task.cancel()
             try:
-                self.hub_connection.stop()
-            except Exception as ex:
-                print(f"Error stopping hub connection: {ex}")
+                await self._signalr_task
+            except asyncio.CancelledError:
+                pass # SignalR task cancelled successfully            
+            except Exception as e:
+                print(f"Error during task cancellation: {e}")
 
+    async def _shutdown_async(self):
+        """Async shutdown helper to cancel all tasks"""     
+
+        # Clear all drain state to prevent semaphore deadlocks
         if self.async_mode:
-            self._executor.shutdown(wait=True)
+            with self._series_lock:
+                self._series_running.clear()
 
-        if self._monitor_thread and self._monitor_thread.is_alive():
-            self._monitor_thread.join(timeout=5)
+        await self._cancel_signalr_task()
+        await asyncio.sleep(0.5)
+        
+        # Cancel all remaining tasks in this event loop except the current one
+        current_task = asyncio.current_task(self._loop)
+        tasks = [
+            t for t in asyncio.all_tasks(self._loop) 
+            if not t.done() and t is not current_task  
+        ]
+        
+        for task in tasks:
+            task.cancel()
+        
+        if tasks: 
+            await asyncio.gather(*tasks, return_exceptions=True)
+        
+        # Final cleanup
+        if self.async_mode:
+            with self._series_lock:
+                self._series_queues.clear()
+            
+    def terminate_hub_connection(self):
+        print("Shutting down DPSHelper...")
+
+        # Shutdown async components
+        if self._loop and self._loop.is_running():
+            future = self._run_in_loop(self._shutdown_async())
+            try:
+                future.result(timeout=10)
+            except Exception as ex:
+                print(f"Error during async shutdown: {ex}")
+
+        # Stop the background event loop
+        if self._loop and self._loop.is_running():
+            self._loop.call_soon_threadsafe(self._loop.stop)
+        
+        # Wait for background thread to finish
+        if self._thread and self._thread.is_alive():
+            self._thread.join(timeout=5)
+            if self._thread.is_alive():
+                print("Warning: Background thread did not stop cleanly")
+        
+        print("DPSHelper shutdown complete")
 
     def subscribe_to_epex_trade_updates(self, handle_data_method: Callable[[object], None]) -> None:
         """
@@ -398,14 +446,16 @@ class DPSHelper:
             handle_data_method `Callable`: A callback function that will be invoked with the received EPEX trade updates.
                 The function should accept one argument, which will be the data received from the EPEX trade updates.
         """
-        if hasattr(self, "epex_trade_call_back") and self.epex_trade_call_back:
-            self.hub_connection.off(EPEX_SUBSCRIPTION_ID)
+        handle_data_method = self._ensure_async(handle_data_method)
+        self._client_initialised.wait()
 
         enact_request_object_epex = [{"Type": "EPEX", "Group": "Trades"}]
 
         self.epex_trade_call_back = handle_data_method
 
-        self._add_subscription(enact_request_object_epex, EPEX_SUBSCRIPTION_ID)
+        self._run_in_loop(
+            self._add_subscription(enact_request_object_epex, EPEX_SUBSCRIPTION_ID)
+        )
 
     def subscribe_to_series_updates(
         self,
@@ -432,6 +482,8 @@ class DPSHelper:
 
             parse_datetimes `bool` (optional): Parse returned DataFrame index to DateTime (UTC). Defaults to False.
         """
+        self._client_initialised.wait()
+        handle_data_method = self._ensure_async(handle_data_method)
         request_details = {"SeriesId": series_id, "CountryId": country_id}
 
         if option_id:
@@ -456,7 +508,10 @@ class DPSHelper:
             )
 
         enact_request_object_series = [request_details]
-        self._add_subscription(enact_request_object_series, subscription_id)
+
+        self._run_in_loop(
+            self._add_subscription(enact_request_object_series, subscription_id)
+        )
 
     def subscribe_to_multiple_series_updates(
         self,
@@ -476,7 +531,9 @@ class DPSHelper:
             parse_datetimes `bool` (optional): Parse returned DataFrame index to DateTime (UTC). Defaults to False.
 
         Note that series, option and country IDs for Enact can be found at https://enact.lcp.energy/externalinstructions.
-        """
+        """      
+        self._client_initialised.wait()
+        handle_data_method = self._ensure_async(handle_data_method)
         join_payload = []
         for series_entry in series_requests:
             series_id = series_entry["seriesId"]
@@ -499,7 +556,9 @@ class DPSHelper:
 
             join_payload.append(series_payload)
 
-        self._add_multi_series_subscription([join_payload], handle_data_method, parse_datetimes)
+        self._run_in_loop(
+            self._add_multi_series_subscription([join_payload], handle_data_method, parse_datetimes)
+        )
 
     def _get_subscription_id(self, series_id: str, country_id: str, option_id: list[str]) -> tuple:
         subscription_id = (series_id, country_id)
@@ -507,19 +566,20 @@ class DPSHelper:
             subscription_id += tuple(option_id)
         return subscription_id
     
+
 if __name__ == "__main__":
-    from manual_tests import *
 
     username = "LcpInternalEnactAccessBaileyHalliday"
     api_key = "28AACrbX79aH"
 
-    #dps_helper = DPSHelper(username, api_key)
-    #epex_test(dps_helper, "epex_old.jsonl")
-    #single_series_test(dps_helper, "single_series_old.jsonl")
-    #notification_test(dps_helper, "notifications_old.jsonl")
-
-    
+    def handle_updates(arg):
+        print(arg)
     dps_helper = DPSHelper(username, api_key)
-    epex_test(dps_helper, "epex_signalrcore.jsonl")
+    dps_helper.subscribe_to_series_updates(handle_updates,"ImbalancePriceRealtime", country_id="Belgium", parse_datetimes=True)
 
+    # keep alive
+    message = None
+    while message != "exit()":
+        message = input(">> ")
 
+    dps_helper.terminate_hub_connection()
