--- lcpdelta_python_package/tests/manual/async_tests/async_mode_logs/test2/pysignalr.jsonl
@@ -1,19 +1,0 @@
-{"timestamp": "2026-01-26T16:08:48.663282+00:00", "data": "                      Gb&RealtimeDemand&none              DateTimeLastUpdated\n2026-01-26T16:05:00Z                 40688.0 2026-01-26 16:08:48.662194+00:00"}
-{"timestamp": "2026-01-26T16:08:58.873282+00:00", "data": "                      Belgium&ImbalancePriceRealtime&none              DateTimeLastUpdated\n2026-01-26T16:04:00Z                               159.01 2026-01-26 16:08:58.871761+00:00\n2026-01-26T16:05:00Z                               153.68 2026-01-26 16:08:58.872370+00:00\n2026-01-26T16:06:00Z                               149.74 2026-01-26 16:08:58.872720+00:00\n2026-01-26T16:07:00Z                               139.76 2026-01-26 16:08:58.873061+00:00"}
-{"timestamp": "2026-01-26T16:09:47.646147+00:00", "data": "                      Gb&RealtimeDemand&none              DateTimeLastUpdated\n2026-01-26T16:06:00Z                 40746.0 2026-01-26 16:09:47.644183+00:00"}
-{"timestamp": "2026-01-26T16:10:16.240321+00:00", "data": "                      Gb&RealtimeFrequency&none              DateTimeLastUpdated\n2026-01-26T16:08:00Z                     50.044 2026-01-26 16:10:16.231350+00:00\n2026-01-26T16:08:15Z                     50.042 2026-01-26 16:10:16.233013+00:00\n2026-01-26T16:08:30Z                     50.031 2026-01-26 16:10:16.235023+00:00\n2026-01-26T16:08:45Z                     50.027 2026-01-26 16:10:16.235827+00:00\n2026-01-26T16:09:00Z                     50.032 2026-01-26 16:10:16.236658+00:00\n2026-01-26T16:09:15Z                     50.040 2026-01-26 16:10:16.237441+00:00\n2026-01-26T16:09:30Z                     50.020 2026-01-26 16:10:16.238181+00:00\n2026-01-26T16:09:45Z                     50.010 2026-01-26 16:10:16.238995+00:00"}
-{"timestamp": "2026-01-26T16:10:51.260605+00:00", "data": "                      Gb&RealtimeDemand&none              DateTimeLastUpdated\n2026-01-26T16:07:00Z                 40792.0 2026-01-26 16:10:51.259763+00:00"}
-{"timestamp": "2026-01-26T16:11:47.775525+00:00", "data": "                      Gb&RealtimeDemand&none              DateTimeLastUpdated\n2026-01-26T16:08:00Z                 40734.0 2026-01-26 16:11:47.770460+00:00"}
-{"timestamp": "2026-01-26T16:12:13.308723+00:00", "data": "                      Gb&RealtimeFrequency&none              DateTimeLastUpdated\n2026-01-26T16:10:00Z                     49.968 2026-01-26 16:12:13.296517+00:00\n2026-01-26T16:10:15Z                     49.958 2026-01-26 16:12:13.298006+00:00\n2026-01-26T16:10:30Z                     49.960 2026-01-26 16:12:13.300058+00:00\n2026-01-26T16:10:45Z                     49.949 2026-01-26 16:12:13.302179+00:00\n2026-01-26T16:11:00Z                     49.962 2026-01-26 16:12:13.305942+00:00\n2026-01-26T16:11:15Z                     49.990 2026-01-26 16:12:13.306744+00:00\n2026-01-26T16:11:30Z                     50.006 2026-01-26 16:12:13.307538+00:00\n2026-01-26T16:11:45Z                     49.971 2026-01-26 16:12:13.308281+00:00"}
-{"timestamp": "2026-01-26T16:12:47.646744+00:00", "data": "                      Gb&RealtimeDemand&none              DateTimeLastUpdated\n2026-01-26T16:09:00Z                 40796.0 2026-01-26 16:12:47.645803+00:00"}
-{"timestamp": "2026-01-26T16:12:58.603984+00:00", "data": "                      Belgium&ImbalancePriceRealtime&none              DateTimeLastUpdated\n2026-01-26T16:08:00Z                               139.76 2026-01-26 16:12:58.602084+00:00\n2026-01-26T16:09:00Z                               146.69 2026-01-26 16:12:58.602823+00:00\n2026-01-26T16:10:00Z                               148.93 2026-01-26 16:12:58.603253+00:00\n2026-01-26T16:11:00Z                               151.45 2026-01-26 16:12:58.603736+00:00"}
-{"timestamp": "2026-01-26T16:13:47.762339+00:00", "data": "                      Gb&RealtimeDemand&none              DateTimeLastUpdated\n2026-01-26T16:11:00Z                 40786.0 2026-01-26 16:13:47.758568+00:00"}
-{"timestamp": "2026-01-26T16:14:44.121579+00:00", "data": "                      Gb&RealtimeFrequency&none              DateTimeLastUpdated\n2026-01-26T16:12:00Z                     49.963 2026-01-26 16:14:44.115034+00:00\n2026-01-26T16:12:15Z                     49.961 2026-01-26 16:14:44.116646+00:00\n2026-01-26T16:12:30Z                     49.952 2026-01-26 16:14:44.117430+00:00\n2026-01-26T16:12:45Z                     49.947 2026-01-26 16:14:44.118154+00:00\n2026-01-26T16:13:00Z                     49.947 2026-01-26 16:14:44.118927+00:00\n2026-01-26T16:13:15Z                     49.952 2026-01-26 16:14:44.119700+00:00\n2026-01-26T16:13:30Z                     49.946 2026-01-26 16:14:44.120429+00:00\n2026-01-26T16:13:45Z                     49.931 2026-01-26 16:14:44.121146+00:00"}
-{"timestamp": "2026-01-26T16:14:47.798394+00:00", "data": "                      Gb&RealtimeDemand&none              DateTimeLastUpdated\n2026-01-26T16:12:00Z                 40787.0 2026-01-26 16:14:47.797776+00:00"}
-{"timestamp": "2026-01-26T16:15:47.967837+00:00", "data": "                      Gb&RealtimeDemand&none              DateTimeLastUpdated\n2026-01-26T16:13:00Z                 40842.0 2026-01-26 16:15:47.967229+00:00"}
-{"timestamp": "2026-01-26T16:16:14.679853+00:00", "data": "                      Gb&RealtimeFrequency&none              DateTimeLastUpdated\n2026-01-26T16:14:00Z                     49.934 2026-01-26 16:16:14.674856+00:00\n2026-01-26T16:14:15Z                     49.942 2026-01-26 16:16:14.675510+00:00\n2026-01-26T16:14:30Z                     49.927 2026-01-26 16:16:14.675855+00:00\n2026-01-26T16:14:45Z                     49.928 2026-01-26 16:16:14.676181+00:00\n2026-01-26T16:15:00Z                     49.943 2026-01-26 16:16:14.676489+00:00\n2026-01-26T16:15:15Z                     49.926 2026-01-26 16:16:14.676807+00:00\n2026-01-26T16:15:30Z                     49.932 2026-01-26 16:16:14.677121+00:00\n2026-01-26T16:15:45Z                     49.941 2026-01-26 16:16:14.677810+00:00"}
-{"timestamp": "2026-01-26T16:16:47.702778+00:00", "data": "                      Gb&RealtimeDemand&none              DateTimeLastUpdated\n2026-01-26T16:14:00Z                 40788.0 2026-01-26 16:16:47.702288+00:00"}
-{"timestamp": "2026-01-26T16:16:58.303257+00:00", "data": "                      Belgium&ImbalancePriceRealtime&none              DateTimeLastUpdated\n2026-01-26T16:12:00Z                               153.92 2026-01-26 16:16:58.301051+00:00\n2026-01-26T16:13:00Z                               155.54 2026-01-26 16:16:58.301869+00:00\n2026-01-26T16:14:00Z                               156.56 2026-01-26 16:16:58.302669+00:00\n2026-01-26T16:15:00Z                               169.51 2026-01-26 16:16:58.302991+00:00"}
-{"timestamp": "2026-01-26T16:17:47.708097+00:00", "data": "                      Gb&RealtimeDemand&none              DateTimeLastUpdated\n2026-01-26T16:14:00Z                 40788.0 2026-01-26 16:17:47.707132+00:00"}
-{"timestamp": "2026-01-26T16:18:22.755215+00:00", "data": "                      Gb&RealtimeFrequency&none              DateTimeLastUpdated\n2026-01-26T16:16:00Z                     49.936 2026-01-26 16:18:22.749037+00:00\n2026-01-26T16:16:15Z                     49.925 2026-01-26 16:18:22.750602+00:00\n2026-01-26T16:16:30Z                     49.898 2026-01-26 16:18:22.751397+00:00\n2026-01-26T16:16:45Z                     49.902 2026-01-26 16:18:22.752118+00:00\n2026-01-26T16:17:00Z                     49.889 2026-01-26 16:18:22.752814+00:00\n2026-01-26T16:17:15Z                     49.886 2026-01-26 16:18:22.753512+00:00\n2026-01-26T16:17:30Z                     49.885 2026-01-26 16:18:22.754200+00:00\n2026-01-26T16:17:45Z                     49.891 2026-01-26 16:18:22.754788+00:00"}
-{"timestamp": "2026-01-26T16:18:47.738333+00:00", "data": "                      Gb&RealtimeDemand&none              DateTimeLastUpdated\n2026-01-26T16:15:00Z                 40826.0 2026-01-26 16:18:47.737246+00:00"}
