--- lcpdelta_python_package/tests/manual/async_tests/async_mode_logs/test2/signalrcore.jsonl
@@ -1,19 +1,0 @@
-{"timestamp": "2026-01-26T16:08:48.663120+00:00", "data": "                      Gb&RealtimeDemand&none              DateTimeLastUpdated\n2026-01-26T16:05:00Z                 40688.0 2026-01-26 16:08:48.659435+00:00"}
-{"timestamp": "2026-01-26T16:08:58.871057+00:00", "data": "                      Belgium&ImbalancePriceRealtime&none              DateTimeLastUpdated\n2026-01-26T16:04:00Z                               159.01 2026-01-26 16:08:58.868657+00:00\n2026-01-26T16:05:00Z                               153.68 2026-01-26 16:08:58.869515+00:00\n2026-01-26T16:06:00Z                               149.74 2026-01-26 16:08:58.869949+00:00\n2026-01-26T16:07:00Z                               139.76 2026-01-26 16:08:58.870664+00:00"}
-{"timestamp": "2026-01-26T16:09:47.646056+00:00", "data": "                      Gb&RealtimeDemand&none              DateTimeLastUpdated\n2026-01-26T16:06:00Z                 40746.0 2026-01-26 16:09:47.645540+00:00"}
-{"timestamp": "2026-01-26T16:10:16.240205+00:00", "data": "                      Gb&RealtimeFrequency&none              DateTimeLastUpdated\n2026-01-26T16:08:00Z                     50.044 2026-01-26 16:10:16.224069+00:00\n2026-01-26T16:08:15Z                     50.042 2026-01-26 16:10:16.226451+00:00\n2026-01-26T16:08:30Z                     50.031 2026-01-26 16:10:16.227641+00:00\n2026-01-26T16:08:45Z                     50.027 2026-01-26 16:10:16.228342+00:00\n2026-01-26T16:09:00Z                     50.032 2026-01-26 16:10:16.228862+00:00\n2026-01-26T16:09:15Z                     50.040 2026-01-26 16:10:16.229395+00:00\n2026-01-26T16:09:30Z                     50.020 2026-01-26 16:10:16.234022+00:00\n2026-01-26T16:09:45Z                     50.010 2026-01-26 16:10:16.239773+00:00"}
-{"timestamp": "2026-01-26T16:10:51.260434+00:00", "data": "                      Gb&RealtimeDemand&none              DateTimeLastUpdated\n2026-01-26T16:07:00Z                 40792.0 2026-01-26 16:10:51.258584+00:00"}
-{"timestamp": "2026-01-26T16:11:47.768898+00:00", "data": "                      Gb&RealtimeDemand&none              DateTimeLastUpdated\n2026-01-26T16:08:00Z                 40734.0 2026-01-26 16:11:47.767737+00:00"}
-{"timestamp": "2026-01-26T16:12:13.305003+00:00", "data": "                      Gb&RealtimeFrequency&none              DateTimeLastUpdated\n2026-01-26T16:10:00Z                     49.968 2026-01-26 16:12:13.293216+00:00\n2026-01-26T16:10:15Z                     49.958 2026-01-26 16:12:13.294541+00:00\n2026-01-26T16:10:30Z                     49.960 2026-01-26 16:12:13.298996+00:00\n2026-01-26T16:10:45Z                     49.949 2026-01-26 16:12:13.301117+00:00\n2026-01-26T16:11:00Z                     49.962 2026-01-26 16:12:13.302854+00:00\n2026-01-26T16:11:15Z                     49.990 2026-01-26 16:12:13.303464+00:00\n2026-01-26T16:11:30Z                     50.006 2026-01-26 16:12:13.303967+00:00\n2026-01-26T16:11:45Z                     49.971 2026-01-26 16:12:13.304597+00:00"}
-{"timestamp": "2026-01-26T16:12:47.644628+00:00", "data": "                      Gb&RealtimeDemand&none              DateTimeLastUpdated\n2026-01-26T16:09:00Z                 40796.0 2026-01-26 16:12:47.643748+00:00"}
-{"timestamp": "2026-01-26T16:12:58.601128+00:00", "data": "                      Belgium&ImbalancePriceRealtime&none              DateTimeLastUpdated\n2026-01-26T16:08:00Z                               139.76 2026-01-26 16:12:58.598654+00:00\n2026-01-26T16:09:00Z                               146.69 2026-01-26 16:12:58.599838+00:00\n2026-01-26T16:10:00Z                               148.93 2026-01-26 16:12:58.600382+00:00\n2026-01-26T16:11:00Z                               151.45 2026-01-26 16:12:58.600859+00:00"}
-{"timestamp": "2026-01-26T16:13:47.762516+00:00", "data": "                      Gb&RealtimeDemand&none              DateTimeLastUpdated\n2026-01-26T16:11:00Z                 40786.0 2026-01-26 16:13:47.761288+00:00"}
-{"timestamp": "2026-01-26T16:14:44.130042+00:00", "data": "                      Gb&RealtimeFrequency&none              DateTimeLastUpdated\n2026-01-26T16:12:00Z                     49.963 2026-01-26 16:14:44.126755+00:00\n2026-01-26T16:12:15Z                     49.961 2026-01-26 16:14:44.127573+00:00\n2026-01-26T16:12:30Z                     49.952 2026-01-26 16:14:44.127945+00:00\n2026-01-26T16:12:45Z                     49.947 2026-01-26 16:14:44.128292+00:00\n2026-01-26T16:13:00Z                     49.947 2026-01-26 16:14:44.128609+00:00\n2026-01-26T16:13:15Z                     49.952 2026-01-26 16:14:44.129039+00:00\n2026-01-26T16:13:30Z                     49.946 2026-01-26 16:14:44.129416+00:00\n2026-01-26T16:13:45Z                     49.931 2026-01-26 16:14:44.129756+00:00"}
-{"timestamp": "2026-01-26T16:14:47.799607+00:00", "data": "                      Gb&RealtimeDemand&none              DateTimeLastUpdated\n2026-01-26T16:12:00Z                 40787.0 2026-01-26 16:14:47.799135+00:00"}
-{"timestamp": "2026-01-26T16:15:47.966266+00:00", "data": "                      Gb&RealtimeDemand&none              DateTimeLastUpdated\n2026-01-26T16:13:00Z                 40842.0 2026-01-26 16:15:47.965362+00:00"}
-{"timestamp": "2026-01-26T16:16:14.674114+00:00", "data": "                      Gb&RealtimeFrequency&none              DateTimeLastUpdated\n2026-01-26T16:14:00Z                     49.934 2026-01-26 16:16:14.671006+00:00\n2026-01-26T16:14:15Z                     49.942 2026-01-26 16:16:14.671872+00:00\n2026-01-26T16:14:30Z                     49.927 2026-01-26 16:16:14.672204+00:00\n2026-01-26T16:14:45Z                     49.928 2026-01-26 16:16:14.672523+00:00\n2026-01-26T16:15:00Z                     49.943 2026-01-26 16:16:14.672994+00:00\n2026-01-26T16:15:15Z                     49.926 2026-01-26 16:16:14.673341+00:00\n2026-01-26T16:15:30Z                     49.932 2026-01-26 16:16:14.673642+00:00\n2026-01-26T16:15:45Z                     49.941 2026-01-26 16:16:14.673928+00:00"}
-{"timestamp": "2026-01-26T16:16:47.701520+00:00", "data": "                      Gb&RealtimeDemand&none              DateTimeLastUpdated\n2026-01-26T16:14:00Z                 40788.0 2026-01-26 16:16:47.700809+00:00"}
-{"timestamp": "2026-01-26T16:16:58.300174+00:00", "data": "                      Belgium&ImbalancePriceRealtime&none              DateTimeLastUpdated\n2026-01-26T16:12:00Z                               153.92 2026-01-26 16:16:58.297347+00:00\n2026-01-26T16:13:00Z                               155.54 2026-01-26 16:16:58.298747+00:00\n2026-01-26T16:14:00Z                               156.56 2026-01-26 16:16:58.299136+00:00\n2026-01-26T16:15:00Z                               169.51 2026-01-26 16:16:58.299735+00:00"}
-{"timestamp": "2026-01-26T16:17:47.696758+00:00", "data": "                      Gb&RealtimeDemand&none              DateTimeLastUpdated\n2026-01-26T16:14:00Z                 40788.0 2026-01-26 16:17:47.696003+00:00"}
-{"timestamp": "2026-01-26T16:18:22.679272+00:00", "data": "                      Gb&RealtimeFrequency&none              DateTimeLastUpdated\n2026-01-26T16:16:00Z                     49.936 2026-01-26 16:18:22.674008+00:00\n2026-01-26T16:16:15Z                     49.925 2026-01-26 16:18:22.675546+00:00\n2026-01-26T16:16:30Z                     49.898 2026-01-26 16:18:22.676332+00:00\n2026-01-26T16:16:45Z                     49.902 2026-01-26 16:18:22.677107+00:00\n2026-01-26T16:17:00Z                     49.889 2026-01-26 16:18:22.677842+00:00\n2026-01-26T16:17:15Z                     49.886 2026-01-26 16:18:22.678408+00:00\n2026-01-26T16:17:30Z                     49.885 2026-01-26 16:18:22.678699+00:00\n2026-01-26T16:17:45Z                     49.891 2026-01-26 16:18:22.679005+00:00"}
-{"timestamp": "2026-01-26T16:18:47.703340+00:00", "data": "                      Gb&RealtimeDemand&none              DateTimeLastUpdated\n2026-01-26T16:15:00Z                 40826.0 2026-01-26 16:18:47.702163+00:00"}
