--- single_series_pysignalr.jsonl
@@ -1,2 +1,0 @@
-{"timestamp": "2026-01-26T15:08:34.207094", "data": "                           Belgium&ImbalancePriceRealtime              DateTimeLastUpdated\nGMT Time                                                                                  \n2026-01-26 00:00:00+00:00                          105.38 2026-01-26 15:04:46.180515+00:00\n2026-01-26 00:01:00+00:00                          106.10 2026-01-26 15:04:46.180515+00:00\n2026-01-26 00:02:00+00:00                          106.10 2026-01-26 15:04:46.180515+00:00\n2026-01-26 00:03:00+00:00                          106.10 2026-01-26 15:04:46.180515+00:00\n2026-01-26 00:04:00+00:00                          115.55 2026-01-26 15:04:46.180515+00:00\n...                                                   ...                              ...\n2026-01-26 14:59:00+00:00                          148.50 2026-01-26 15:04:46.180515+00:00\n2026-01-26 15:04:00+00:00                          151.89 2026-01-26 15:08:34.202154+00:00\n2026-01-26 15:05:00+00:00                          152.30 2026-01-26 15:08:34.204328+00:00\n2026-01-26 15:06:00+00:00                          152.72 2026-01-26 15:08:34.205690+00:00\n2026-01-26 15:07:00+00:00                          152.44 2026-01-26 15:08:34.206801+00:00\n\n[900 rows x 2 columns]"}
-{"timestamp": "2026-01-26T15:32:34.085554", "data": "                           Belgium&ImbalancePriceRealtime              DateTimeLastUpdated\nGMT Time                                                                                  \n2026-01-26 00:00:00+00:00                          105.38 2026-01-26 15:32:21.274906+00:00\n2026-01-26 00:01:00+00:00                          106.10 2026-01-26 15:32:21.274906+00:00\n2026-01-26 00:02:00+00:00                          106.10 2026-01-26 15:32:21.274906+00:00\n2026-01-26 00:03:00+00:00                          106.10 2026-01-26 15:32:21.274906+00:00\n2026-01-26 00:04:00+00:00                          115.55 2026-01-26 15:32:21.274906+00:00\n...                                                   ...                              ...\n2026-01-26 15:27:00+00:00                          156.69 2026-01-26 15:32:21.274906+00:00\n2026-01-26 15:28:00+00:00                          156.81 2026-01-26 15:32:34.082550+00:00\n2026-01-26 15:29:00+00:00                          157.09 2026-01-26 15:32:34.083675+00:00\n2026-01-26 15:30:00+00:00                          268.52 2026-01-26 15:32:34.084490+00:00\n2026-01-26 15:31:00+00:00                          283.71 2026-01-26 15:32:34.085254+00:00\n\n[928 rows x 2 columns]"}
